Consolidate dispatcher Socket Mode test doubles
The authorization stub and socket response recorder were defined
identically in test_dispatch.py and test_approval_actions.py despite a
shared tests/conftest.py.

- Move _authorize and FakeSocketClient into tests/conftest.py
- Import them into both modules via the relative-conftest pattern the
  worker sandbox tests already use under importlib mode
- No dispatcher behavior change; the focused tests keep their assertions

Closes #637
