+++ apps/dispatcher/tests/conftest.py
@@ -2,13 +2,39 @@
 compose stack (per repo test discipline: never mock Valkey). Only the Slack Web
 API and socket transport are faked."""
 
+import logging
 import os
 import uuid
 from collections.abc import Iterator
+from typing import Any
 
 import pytest
 import redis
 from agentos_dispatcher.config import DispatcherConfig
+from slack_bolt.authorization import AuthorizeResult
+
+
+def _authorize(**_kwargs: Any) -> AuthorizeResult:
+    """Shared authorization stub: Bolt's ``authorize`` callback resolved to a
+    fixed bot identity, so Socket Mode tests skip the real auth.test call."""
+    return AuthorizeResult(
+        enterprise_id=None,
+        team_id="T1",
+        bot_token="xoxb-test",
+        bot_id="B1",
+        bot_user_id="U0BOT",
+    )
+
+
+class FakeSocketClient:
+    """Captures the envelope acks Bolt sends back over the socket."""
+
+    def __init__(self) -> None:
+        self.logger = logging.getLogger("fake-socket")
+        self.acked_envelope_ids: list[str] = []
+
+    def send_socket_mode_response(self, response: Any) -> None:
+        self.acked_envelope_ids.append(response.envelope_id)
 
 # Compose defaults (compose.dev.yaml maps Valkey to host port 26379, password
 # valkeypass). Overridable for CI via TEST_VALKEY_* env vars.
